% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wunc.R
\name{calculate_commitment}
\alias{calculate_commitment}
\title{Calculate the commitment time series according to Freelon et al.}
\usage{
calculate_commitment(input, after_days = 3)
}
\arguments{
\item{input}{a data frame, either from rtweet::search_tweets or any one with the three columns meaning user id, timestamp and content.}

\item{after_days}{a number to determine what consistents a repeat participation. Default to 3 days (Freelon et al.)}
}
\value{
A tibble with date and the 'unity' metric.
}
\description{
This function calculates the 'commitment' time series from tweets according to Freelon et al. (2019)
}
\references{
Freelon, D., McIlwain, C., & Clark, M. (2018). Quantifying the power and consequences of social media protest. New Media & Society, 20(3), 990-1011.
}
