% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wunc.R
\name{calculate_unity}
\alias{calculate_unity}
\title{Calculate the unity time series according to Freelon et al.}
\usage{
calculate_unity(input)
}
\arguments{
\item{input}{a data frame, either from rtweet::search_tweets or any one with the three columns meaning user id, timestamp and content.}
}
\value{
A tibble with date and the 'unity' metric.
}
\description{
This function calculates the 'unity' time series from tweets according to Freelon et al. (2019)
}
\references{
Freelon, D., McIlwain, C., & Clark, M. (2018). Quantifying the power and consequences of social media protest. New Media & Society, 20(3), 990-1011.
}
